+++ src/bloqade/noise/native/_wrappers.py
@@ -35,7 +35,7 @@ def pauli_channel(
 @wraps(native.CZPauliChannel)
 def cz_pauli_channel(
     ctrls: ilist.IList[Qubit, Any] | list,
-    qarg2: ilist.IList[Qubit, Any] | list,
+    qargs: ilist.IList[Qubit, Any] | list,
     *,
     px_ctrl: float,
     py_ctrl: float,
